CI: Generate bundled images with specific version tag.
Previously, no docker image tag was specified, which means that "latest" was used. This can _theoretically_ lead to drift between the version that we're building for and the version of the server bundled in the image.

Now, the specific version of the server we're building for is used to select the specific `gotify/server` image. This also means past runs are reproducable.